move save function to helpers
